Rename model to match database table
