+++ src/persistence/biome.js
@@ -1,4 +1,4 @@
-module.exports = (bookshelfConn, { BiomeByEA }) => {
+module.exports = (bookshelfConn, { GeoEABiome }) => {
   const { knex } = bookshelfConn;
 
   return {
@@ -10,7 +10,7 @@ module.exports = (bookshelfConn, { BiomeByEA }) => {
      * @return {Array} JSON Objects
      */
     findBiomeByEA: envAuthority => (
-      BiomeByEA
+      GeoEABiome
         .where('id_ea', envAuthority)
         .fetchAll({
           columns: ['gid', 'name_biome', 'id_ea', knex.raw('ST_AsGeoJSON(geom) as "geomGeoJSON"')],
